requirements.txt
cython
hidapi
